+++ backend/apps/contracts/admin/contractpayment_admin.py
@@ -10,6 +10,7 @@ from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 
 from apps.contracts.models import ContractPayment, Invoice, InvoiceStatus
+from simple_history.admin import SimpleHistoryAdmin
 
 if TYPE_CHECKING:
     BaseModelAdmin = admin.ModelAdmin